requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
pandas==2.1.3
numpy==1.26.2
psycopg2-binary==2.9.9
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
apache-airflow==2.7.3
